refactor confirm dialog
